added character and item tests

+++ src/main/java/items/IAttack.java
@@ -1,5 +1,5 @@
 package items;
 
 public interface IAttack {
-    double getDamage();
+    int getDamage();
 }
